+++ tests/unit/core/fetcher/test_bot_detection.py
@@ -2,8 +2,8 @@
 
 Covers every decision branch in:
   - HTMLFetcher._check_for_bot_detection  (base.py)
-  - SimpleFetcher.fetch  — logger.warning emitted on detection  (simple.py)
-  - Pipeline._handle_bot_detection  — logger.warning + observability.warning emitted  (pipeline.py)
+  - SimpleFetcher.fetch  -- logger.warning emitted on detection  (simple.py)
+  - Pipeline._handle_bot_detection  -- logger.warning + observability.warning emitted  (pipeline/extraction.py)
 """
 
 import pytest
@@ -12,10 +12,6 @@ from yosoi.core.fetcher.base import HTMLFetcher
 from yosoi.core.fetcher.simple import SimpleFetcher
 from yosoi.core.pipeline import Pipeline
 from yosoi.utils.exceptions import BotDetectionError
-
-# ---------------------------------------------------------------------------
-# Minimal concrete HTMLFetcher so we can call the mixin method directly
-# ---------------------------------------------------------------------------
 
 VALID_HTML = '<html><body>' + 'x' * 200 + '</body></html>'
 
@@ -30,21 +26,11 @@ def fetcher() -> _ConcreteHTMLFetcher:
     return _ConcreteHTMLFetcher()
 
 
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
 def _make_pipeline_stub(mocker):
     stub = Pipeline.__new__(Pipeline)
     stub.console = mocker.MagicMock()
     stub.logger = mocker.MagicMock()
     return stub
-
-
-# ===========================================================================
-# _check_for_bot_detection — HTML-too-short guard
-# ===========================================================================
 
 
 def test_short_html_is_blocked(fetcher):
@@ -65,14 +51,8 @@ def test_exactly_99_chars_is_blocked(fetcher):
 
 
 def test_exactly_100_chars_is_not_short_blocked(fetcher):
-    # 100-char body with no bad patterns → not blocked on 200
     blocked, _ = fetcher._check_for_bot_detection('x' * 100, 200)
     assert blocked is False
-
-
-# ===========================================================================
-# _check_for_bot_detection — hard-block status codes (403 / 429 / 503)
-# ===========================================================================
 
 
 @pytest.mark.parametrize('status_code', [403, 429, 503])
@@ -112,11 +92,6 @@ def test_hard_block_enriched_with_all_cf_headers(fetcher, status_code):
     assert 'CF-Ray: xyz-AMS' in indicators
 
 
-# ===========================================================================
-# _check_for_bot_detection — cf-mitigated header (any status)
-# ===========================================================================
-
-
 @pytest.mark.parametrize('status_code', [200, 301, 404, 500])
 def test_cf_mitigated_header_blocks_on_any_status(fetcher, status_code):
     headers = {'cf-mitigated': 'challenge'}
@@ -132,11 +107,6 @@ def test_cf_mitigated_also_enriched_with_cf_ray(fetcher):
     assert 'CF-Ray: ray42-CDG' in indicators
 
 
-# ===========================================================================
-# _check_for_bot_detection — retry-after header on >=400 (non-hard-block)
-# ===========================================================================
-
-
 def test_retry_after_on_400_adds_rate_limit_indicator(fetcher):
     headers = {'Retry-After': '60'}
     _, indicators = fetcher._check_for_bot_detection(VALID_HTML, 400, headers)
@@ -147,11 +117,6 @@ def test_retry_after_on_200_does_not_add_rate_limit_indicator(fetcher):
     headers = {'Retry-After': '60'}
     _, indicators = fetcher._check_for_bot_detection(VALID_HTML, 200, headers)
     assert 'Rate limited (Retry-After header)' not in indicators
-
-
-# ===========================================================================
-# _check_for_bot_detection — 200 OK body patterns (Cloudflare + CAPTCHA)
-# ===========================================================================
 
 
 @pytest.mark.parametrize(
@@ -192,7 +157,6 @@ def test_200_patterns_are_case_insensitive(fetcher):
 
 
 def test_200_pattern_only_checked_in_first_3000_chars(fetcher):
-    # Pattern buried well past 3000 chars should NOT trigger
     html = 'x' * 100 + 'x' * 3000 + 'challenge-platform'
     blocked, _ = fetcher._check_for_bot_detection(html, 200)
     assert blocked is False
@@ -203,11 +167,6 @@ def test_200_multiple_patterns_all_reported(fetcher):
     _, indicators = fetcher._check_for_bot_detection(html, 200)
     assert 'Google reCAPTCHA' in indicators
     assert 'Cloudflare Turnstile' in indicators
-
-
-# ===========================================================================
-# _check_for_bot_detection — >=400 body patterns (legacy + shared)
-# ===========================================================================
 
 
 @pytest.mark.parametrize(
@@ -238,21 +197,11 @@ def test_non_hard_block_500_with_body_pattern_blocked(fetcher):
     assert blocked is True
 
 
-# ===========================================================================
-# _check_for_bot_detection — 2xx/3xx non-200 statuses without body patterns
-# ===========================================================================
-
-
 @pytest.mark.parametrize('status_code', [201, 204, 301, 302])
 def test_non_200_success_or_redirect_not_blocked(fetcher, status_code):
     blocked, indicators = fetcher._check_for_bot_detection(VALID_HTML, status_code)
     assert blocked is False
     assert indicators == []
-
-
-# ===========================================================================
-# _check_for_bot_detection — headers=None is backwards-compatible
-# ===========================================================================
 
 
 def test_no_headers_argument_does_not_crash(fetcher):
@@ -265,17 +214,11 @@ def test_empty_headers_dict_does_not_crash(fetcher):
     assert blocked is False
 
 
-# ===========================================================================
-# SimpleFetcher.fetch — logger.warning emitted on bot detection
-# ===========================================================================
-
-
 @pytest.mark.asyncio
 async def test_simple_fetcher_logs_warning_on_bot_detection(mocker):
     fetcher_instance = SimpleFetcher(use_session=False)
     fetcher_instance.logger = mocker.MagicMock()
 
-    # Craft a response that triggers bot detection (403)
     mock_response = mocker.MagicMock()
     mock_response.status_code = 403
     mock_response.text = VALID_HTML
@@ -336,38 +279,42 @@ async def test_simple_fetcher_no_warning_on_clean_200(mocker):
 
 
 # ===========================================================================
-# Pipeline._handle_bot_detection — logger.warning + observability.warning emitted
+# Pipeline._handle_bot_detection
+# _handle_bot_detection lives in pipeline/extraction.py — patch that module's
+# logger and observability bindings, not yosoi.core.pipeline.base.
 # ===========================================================================
 
 
 def test_handle_bot_detection_calls_logger_warning(mocker):
     stub = _make_pipeline_stub(mocker)
-    mocker.patch('yosoi.core.pipeline.observability')
+    mocker.patch('yosoi.core.pipeline.extraction.observability')
+    mock_logger = mocker.patch('yosoi.core.pipeline.extraction.logger')
 
     err = BotDetectionError(url='https://x.com', status_code=403, indicators=['HTTP 403', 'CF-Ray: abc'])
     Pipeline._handle_bot_detection(stub, err, attempt=1, max_retries=3)
 
-    stub.logger.warning.assert_called_once()
-    call_args = str(stub.logger.warning.call_args)
+    mock_logger.warning.assert_called_once()
+    call_args = str(mock_logger.warning.call_args)
     assert 'https://x.com' in call_args
     assert '403' in call_args
 
 
 def test_handle_bot_detection_logger_warning_includes_attempt_info(mocker):
     stub = _make_pipeline_stub(mocker)
-    mocker.patch('yosoi.core.pipeline.observability')
+    mocker.patch('yosoi.core.pipeline.extraction.observability')
+    mock_logger = mocker.patch('yosoi.core.pipeline.extraction.logger')
 
     err = BotDetectionError(url='https://x.com', status_code=429, indicators=['HTTP 429'])
     Pipeline._handle_bot_detection(stub, err, attempt=2, max_retries=3)
 
-    call_args = str(stub.logger.warning.call_args)
+    call_args = str(mock_logger.warning.call_args)
     assert '2' in call_args
     assert '3' in call_args
 
 
 def test_handle_bot_detection_calls_obs_warning(mocker):
     stub = _make_pipeline_stub(mocker)
-    mock_obs = mocker.patch('yosoi.core.pipeline.observability')
+    mock_obs = mocker.patch('yosoi.core.pipeline.extraction.observability')
 
     err = BotDetectionError(url='https://x.com', status_code=503, indicators=['HTTP 503'])
     Pipeline._handle_bot_detection(stub, err, attempt=1, max_retries=2)
@@ -382,7 +329,7 @@ def test_handle_bot_detection_calls_obs_warning(mocker):
 
 def test_handle_bot_detection_obs_includes_indicators(mocker):
     stub = _make_pipeline_stub(mocker)
-    mock_obs = mocker.patch('yosoi.core.pipeline.observability')
+    mock_obs = mocker.patch('yosoi.core.pipeline.extraction.observability')
 
     indicators = ['HTTP 403', 'Cloudflare server', 'CF-Ray: ray1-LHR']
     err = BotDetectionError(url='https://x.com', status_code=403, indicators=indicators)
@@ -394,7 +341,7 @@ def test_handle_bot_detection_obs_includes_indicators(mocker):
 
 def test_handle_bot_detection_still_prints_to_console(mocker):
     stub = _make_pipeline_stub(mocker)
-    mocker.patch('yosoi.core.pipeline.observability')
+    mocker.patch('yosoi.core.pipeline.extraction.observability')
 
     err = BotDetectionError(url='https://x.com', status_code=403, indicators=['HTTP 403'])
     Pipeline._handle_bot_detection(stub, err, attempt=1, max_retries=2)
@@ -404,7 +351,7 @@ def test_handle_bot_detection_still_prints_to_console(mocker):
 
 def test_handle_bot_detection_abort_message_when_exhausted(mocker):
     stub = _make_pipeline_stub(mocker)
-    mocker.patch('yosoi.core.pipeline.observability')
+    mocker.patch('yosoi.core.pipeline.extraction.observability')
 
     err = BotDetectionError(url='https://x.com', status_code=403, indicators=['HTTP 403'])
     Pipeline._handle_bot_detection(stub, err, attempt=2, max_retries=2)
@@ -415,7 +362,7 @@ def test_handle_bot_detection_abort_message_when_exhausted(mocker):
 
 def test_handle_bot_detection_no_abort_message_when_retries_remain(mocker):
     stub = _make_pipeline_stub(mocker)
-    mocker.patch('yosoi.core.pipeline.observability')
+    mocker.patch('yosoi.core.pipeline.extraction.observability')
 
     err = BotDetectionError(url='https://x.com', status_code=403, indicators=['HTTP 403'])
     Pipeline._handle_bot_detection(stub, err, attempt=1, max_retries=3)
